+++ lab06/main.c
@@ -3,14 +3,14 @@
 #include <stdlib.h>
 #include <math.h>
 
-double mzero (double xl, double xp, double eps, int metoda, int *kroki);
+float mzero (float xl, float xp, float eps, int metoda, int *kroki);
 int main (int wybor, char ** argv)
 {
   if (wybor < 2) {printf("Nie wybrano metody poszukiwania miejsca zerowego.\nPodaj argument wywołania\n 0 - metoda bisekcji \n 1 - metoda stycznych\n");
       return -1;}
   else {
 char u_xl[10], u_xp[10], u_eps[10]; /* stringi zachowujące input użytkownika */
-double xl, xp, eps;
+float xl, xp, eps;
 char *err_xl, *err_xp, *err_eps; /*zmienne pomocnicze, pozwalające wykryć błąd */
 int metoda, kroki;
     printf("\nSzukacz miejsc zerowych funkcji\n");
